Imported motioner project source.
The project is for creating motion files that can be loaded into
VastSpace program.  It has codes that should be shared among the
main source.

The wtltest subdirectory reimplements the GUI with Windows Template
Library.  You'll need WTL 8.1 installed in a subdirectory of parent
directory of the repository to compile the project wtltest.vcproj.