surface: Add roll and pitch indicator to F15's HUD.
We had a hard time figuring out how to convert a quaternion to an Euler
angle set, i.e. pitch, yaw and roll.
There are tons of information out there on the Internet, but the problem
is that all the examples use distinct conventions for the Euler angles.
So we had to experiment with lots of different candidates to find
the answer.  This effort resulted in a script file for calculating
roll, namely roll.nut.

Finally, we have found the answer with analytic means inspired by:
	http://graphics.wikia.com/wiki/Conversion_between_quaternions_and_Euler_angles