surface: Add readouts of pitch, yaw and roll to F15's HUD.
This is of course not present in the real vehicle, but it's easier to
inspect if something wrong is happening.

Also note that we use lambda expression, feature introduced in C++11, for
the first time in this project.  This feature helps us keeping locality of
functions as high as possible.