surface: Make F15 capable of shooting missiles.
The Missile class is designed for space, so it doesn't behave
realistically in surface.  Specifically, it doesn't take effect of
gravity.
We should make a dedicated class, say AIM-9, for the real equipment.
Also we need a model for display.

For the time being, we reuse the logic of space missiles and it responds
well.

Also demo7.nut is made to restore health of the Entities with reset().