Made the GLWchart to plot time taken by various renderings.
Specifically, F15 looks to take too much time to render in certain
runtime environment (Core i7 860 with 12GB RAM) with even a single
instance.  It may be something to do with memory management.  Probably
we should make out something to enhance the performance, but for the
time being, measuring the current phenomena is a good starting point.