surface: Tank shoots only if it has clear line of sight to the target.
Deterimining if we have a clear line of sight is done by using traceHit().
Tracing ray for long range is costly process, so we do it only once
in a while to suppress CPU load.  This check interval is randomized among
Tank instances to prevent concentration of load on a specific period.

We also fixed TIN::traceHit() accessing out of range of the grid.