surface: Add M3Truck class.
It can run just like Tank with a little bit faster speed, but its
armaments are very weak.
Currently, it will enclose to enemy blindly, which makes it a good
prey for Tanks, but it's good starting point for debugging.

An intermediate abstract class LandVehicle is added in this commit
to share common behavior among Tank and M3Truck.