surface: Make M3Truck's mouted gun automatically tracks target.
Also revised gun parameters (such as rate of fire and damage) to get
it closer to real gun.

M3Truck checks line of sight to determine if it can shoot the gun
without terrain got in the way, just like Tank's main gun.
It costs fair amout of CPU time to do it for long path, just like
the case of Tank.