cpplib: Renamed Mat3::vp3() to Mat3::vp().
The reason is that it's obvious that the vector element count is 3.
Appending 3 to the tail makes the matrix class compatible with Mat4,
but it's not consistent with clib's MATVP() macro.
After all, we have decided that making mathematically natural naming
convention is better than treating augmented matrix
(http://en.wikipedia.org/wiki/Augmented_matrix)
a common thing, which is really special usage of 4-D matrix applied
to OpenGL graphics.

Also fixed some application codes that using the method.