Add forward declaration header file for Model structure.
We have been added forward declaration of classes and structs as
incomplete type declarations just before their uses, but that method
introduces base type dependence whether they're classes or structs.
In other words, we cannot easily convert a struct to a class or vice
versa if it's forward declared in place.

This kind of dependence to type information should be eliminated as
possible.  A forward declaration header file is a way to do this,
although using them increases number of header files excessively.

As a starting point, I chose Model structure.