surface: Add HydraRocketLauncher class with model for Apache.
It utilizes ArmBase and hardpoint_static mechanism, originally for
space ships, to enable customization of the set of equipments.
I feel a bit unconfortable to reuse the mechanism, because the
existing mechanism assumes that ArmBase derived objects aim at
targets on their own, while they just carry ammo and shoot on
command in the case of Apache (or other possible attack helicopters).
But I accepted it because it just works.