Remove EXPORT macro from WeakPtr's declaration.
WeakPtr is a class template with all member functions defined as inlines,
so we should not add EXPORT macro to its declaration, or the extension
modules end up with linker errors in Visual C++ compiler.
I didn't investigate the issue on Linux gcc, but I suspect gcc is OK in
either way.