Revived equality operators for WeakPtr.
They are disabled because they tend to mess up with complex conditional
statements due to template type inference, but adjusting the logic a bit
made them compile fine in the whole project.

Defining those operators explicitly hopefully avoids unnecessary WeakPtr
object creation in the comparison expression.

To be honest, there's no intrinsic reason to fix this issue now.
The practical reason is that Observer.h is modified for another issue,
which make the whole project recompile, which in turn makes me to think
that it's a good timing to do this.