surface: Experimentally generated gun bullets from Squirrel scripts for A-10.
In this way, we have almost all the flexibility of scripts for fixed
guns too.  The problem is that we should advance the physical simulation
time in subframe precision to make it realistic if the gun's rate of fire
is relatively high, which is not achieved in this commit without help of
native codes.