Add RandomSequencePtr class accessible through CoordSys from Squirrel.
The class is merely a pointer to RandomSequence object.  It's necessary
to be a pointer because pseudo random number query is mutable operation
which should affect internal state vector of the RandomSequence object.

RandomSequence is member of WarField, not really a member of CoordSys,
but it's the same thing from Squirrel script's point of view.

There's a known vulnerability in this scheme that accessing freed memory
is possible through holding RandomSequencePtr object after its containing
object (WarField) is destroyed.  This case should be extremely rare but
certainly possible.  We don't take care of this issue right now, but
adding weak pointer scheme or something is necessary to work around
this correctly.