Make Vec3d's elements accessible through indexing in Squirrel scripts.
It means you can write like vec[0] to access the first element of a
vector.  This simple syntax have been not supported because you can
write basic linear algebra without it.  But now that we are moving much
logic to scripts, we'd like to use the syntax.

The syntax is also free from 'updating temporary object' pitfall by
compound assignment operators, such as +=, -= and *=.