Make Teline3 more object oriented.
We made the struct be virtual and defined various forms of Teline3
by deriving the base class.  Currently only LineTeline3 and
CylinderTeline3 are tried.  CylinderTelin3 is not even tested.
But I'm confident that it'll work on all the variants.

The drawback of new object oriented mechanism is that we cannot know
the size of required buffer to hold all possible forms if we disclose
the base class definition to allow user derived forms.
I'm still thinking about how to deal with this problem.