surface: Workaround for the bug Launchers not working since ffadddc.
The commit ffadddc made Entity::sq_pushobj() to use classname() of
the Entity in question, which is not necessarily defined as a symbol in
the root table of Squirrel VM.  That change was necessary to use
functions defined in derived classes, but it broke working Launcher
class branch for aerial force armaments.  Currently, A10 suffers from the
issue that it can only shoot the main gun, i.e. Avenger.

It's a tough job to fundamentally fix the problem, but instead we
patched surface module to get Launcher-derived classes working for
the time being.