Add brand new Mesh class, which has some compatibility to suf_t in clib.
The clib library is fairly mature and stable, which makes me don't feel
like touching it.  The suf_t aspect of the library is used by other
applications, so we cannot just extend it without accepting risk of
compatibility failure.  It's also written in plain C, which renders
it very hard to integrate to a well object-oriented C++ application
or extend like Model and Motion.

So we decided to make a branch to alter suf_t with brand new class.
Because the library has much of arcane techniques that should take
time to port, we should start off with merely a dead copy of suf_t
and gradually make it objective.

I have tested that the new classes, Mesh and MeshTex, work with
textures and shadow mapping in VC2012 build.