Mesh: Make use of VBO (Vertex Buffer Objects) to enhance performance.
It's expected to be effective against complex objects with uniform material.
We don't have many of such kind of models, but the effect is valuable
if shadow mapping is enabled, because shadow mapping renders the scene
multiple times in a frame.

From this commit later on, Mesh class need to be newed instead of malloc'ed.
This is because a STL container is added to the class.  But it won't be a
problem if the user is object oriented.