surface: Refactored WarMap's traceHit method interface.
It used to be similar to Entity::traceHit(), but I feel it's better
to write it in more object oriented manner.  Namely, input values
and output values are structurized to be less cluttering and
more extensive.  This is effective for traceHit() because it tend
to have many arguments for both inputs and outputs.