surface: Even more improved Aerial autopilot for landing.
Now it can land to an Airport precisely, given that the starting position
is good for landing.

Added aerialLandingSpoiler function to return desired spoiler status
from the autopilot script.

Also added sqGet entries to Airport for querying ILS signals from a
Squirrel script.

The scripted autopilot has proven itself useful for rapid prototyping with
its ability to reload on the fly (during the game is running).
We can run the extremely fast 2-step cycle, that is, tweaking the autopilot
by modifying the script and reloading in the game, to brush up the logic.