Move aerialLanding definition code to surface module.
This is done because the codes are only necessary if surface module is
loaded.  You won't need to dofile("aerial.nut") for all possible cases.
Now it's loaded when an Aerial vehicle is created.

We have defined it in the main executable's init.nut because we thought
that defining a function as an entry of the root table from initialization
scripts is not possible, which was correct in some extent, but exploiting
Squirrel language features made it possible.

We merely moved aerial.nut to a subdirectory of surface module but not
directly embedded its contents to surface/models/Aerial.nut for two
reasons.
  1. Squirrel scripts are hard to fix syntax errors if they have many
  lines in a file.
  2. Git is better at detecting file moves than content moves.