surface: Change the way to store script controlled values for Aerial.
In aerialLanding.nut, we have been used a global table with Entity IDs
as indices to store script relevant parameters for each object.
It has been worked, but has a problem that garbage entries will remain
in the table after deleting an associated object.  Adding delete hook to
clear the table would be a way, but we'd like to take more direct solution;
Aerial instance has an arbitrary Squirrel object as its parameter.
This method is proven to work in HellfireProc().