Add mechanism to load popup menu items from Squirrel scripts to Entity
Each time the player clicks right mouse button on the screen,
Entity::popupMenu() tries to load a list of popup menu items from a
executed result of a Squirrel function.  This way, we can fully customize
the contents of the menu without recompiling.  In addition, we have
opportunity to dynamically tweak the menu items.

As a demonstration, Sceptor utilizes this mechanism.

The overhead of executing scripted function is negligible, because it's so
rare that the method is invoked.