Add sqserial_findobj() to prevent duplicate push to Squirrel VM
Previously, we have been pushing Serializable objects without concern
about whether the object is already been instantiated in the VM,
and it has been working mostly fine.  Recently we noticed this
mechanism's shortcomings as we add various scripts.  For instance,
If we evaluate "player.cs==player.cs", it returns false.
So we decided to insert a table to prevent duplicate instances for
actually the same object in the game to be pushed to Squirrel VM.

The table has Serializable IDs as keys and instances defined in
Squirrel as values.  If we were to push a Serializable object,
its ID is first searched in the table to find if an instance of
the object is already present in the VM.  If so, we can push a
reference to the already present one instead of creating new one.

We should remove deleted object entries in the table with the
WeakPtr's release event.