audio: Import sound mixing library from clib
We want to convert the library to C++ and also customize it for our
project's need, so we copied it instead of tweaking existing library.
Another important reason is that exporting functions as DLL symbols
in a static library is counter-intuitive in Visual C++.  It's not
impossible, but we need to specify all functions to be exported in
a .def file.  If it comes to class methods, we need to write down
decorated names.

The sound mixing is performed in a dedicated thread so that the main
thread won't block the sound and vice versa.

I think there are bugs regarding positioning sound source in 3-D
space, which should be fixed with extensive tests.

We should really use common libraries such as OpenAL or DirectSound
for stability and portablility, but implementing one by ourselves
would be fun.