surface: Make Tank gun sound update position after shot
The tank gun sound is heavy and long.  The vehicle can move away from
the point it shot the gun, in which case the listener in the vehicle
(or chasing as an external camera) hear the sound source drifts away
from the vehicle.

In reality, the sound source shouldn't move.  This is rough
approximation of the phenomena.  Sound waves reflect about the
surrounding objects, such as ground, wall, other vehicles or
structures.  This multi-pass echoes accumulate together to make
apparently direction-less sound.  Probably we should add a property
like direction-less-ness (or whatever) to 3-D sound sources to
make large sounds feel real.