audio: Import libogg, the container format library for Vorbis
This commit is the first half of Ogg Vorbis support.

The ogg library (libogg) apparently managed with Subversion, but
there's a Git mirror at git://git.xiph.org/mirrors/ogg.git, so we
can subtree merge the library (as we did with clib and cpplib).
I've actually tried that and it worked without a problem, but I
don't think it's worth keeping all the history information of the
library inside our project.  The library is fairly stable and we
can expect that we won't need to frequently update the library.
We just copy the latest source tree under libogg directory,
just like zlib, jpeg, lpng and squirrel3.