Prevent errors on referencing MTurret from a script
ArmBases (and in turn MTurrets) are always created with mounted platform,
which makes direct creation of those Entities are prohibited from a
Squirrel script.
We have not defined EntitiyRegister<MTurret> to achieve this effect,
but it renders scripts unable to access those Entities at all.
The correct way is to define the EntityRegister but override creation
routine by instantiating class template.