Separate EntityRegister implementation to dedicated header
The EntityRegister mechanism is only referenced by certain sources
(namely Entity-derived class implementations), so exposing it to
all the sources that include Entity.h is not optimal.
Now the dedicated header EntityRegister.h contains the implementation
of EntityRegister class template.  This new header file requires
inserting an include directive to every Entity descendant sources.
But it will reduce fake dependency between sources and headers.

Also inserted an intermediate class hierarchy between EntityStatic
and EntityRegister; EntityRegisterNC is introduced.
The new class template is used for Entities that cannot be created
by the user scripts.  Typical example of those Entities are turrets
which require mounted platform prior to create one.  Another example
is TorusStationEntity which is created by TorusStation (CoordSys
descendant), so the user script should not be able to create it.