Fix debri drawing callback to load the MQO file directly
Previously, we needed precompiled .bin files of debris or the game
crashed when it tries to draw a debri.  Now we load the model from
the MQO file directly and properly check the state before drawing
to prevent crashing.

Also, Teline3DrawData has a new member variable to pass a user provided
pointer around.  This is necessary to make the debris draw without
textures in shaders.