Make MTurret descendants use hitRadius for culling
The culling estimation radius is hard-coded, but we should use some
geometry-relevant value to estimate size of the model.
Hit radius is not the exact value for this purpose, but it should
be nearly the best value.  If you see models being culled
inappropriately, probably the hit radius should be revisited,
because bullet hit checks are done with the hit radius also.