Add buildBodyByHitboxes() to ModelEntity to share for multiple Entities
The btShape and btRigidBody construction with a list of hitboxes are
so common that it appears dozens of times.

We have not shared the code because many classes share a shape among
its instances, but we shouldn't repeat the same codes just for this
reason.

The original source is moved from Destroyer::buildBody().