names: Get rid of lambda expression in doubt about performance
C++'s lambda expressions are powerful tool that shortens many lines of
codes, but if it's making the program slow, we should avoid using it.
We are performing real-time, time-critical task to generate random names.
We should complete the task in a instant or process in parallel to
prevent freezes in the main game.

Now, it isn't making noticeable difference after getting rid of them.