names: Add a dedicated window class for selecting warp destination
GLWmenu could not handle so many warp destinations that can be created
by random star generation.  The new class, GLWwarp, has no scroll bars
yet, but you can sort items by name or distance.

GLWwarp could be defined in a dedicated header file, but for the moment,
we do not need to use the window anywhere outside of Warpable-draw.cpp.