names: Move warp destination window definition to initClient.nut
The scripted version can do everything the hard-coded version could,
so why we would stick to the latter?

Probably execution speed is slightly better when we use the hard-coded
version, but the most time-consuming part is the graphics rendering.
After all, choosing scripted or hard-coded here does not impact the
performance much.