Silence warnings involving exporting classes from DLL containing STL
The warnings are so verbose that other warnings have been drown out.
Usually, we should not ignore a warning easily.  Nevertheless, we have
no effective way to avoid this warning; we cannot add __declspec(export)
to the STL classes sources.  We cannot quit using STL containers in
exported classes easily.  We know that it's working nicely without
addressing the warnings for years.  We don't get the warnings in gcc.

Also disabling these warnings make building faster.