Try hard to make creation and deletion of font as few as possible
There's a strange issue that running the program occasionally make
the Windows OS's fonts corrupt.  Once happen, it won't restore
until system reboot.  It's quite difficult to debug because the
phenomenon occurs once in dozens of times running the program.

We have encountered the same symptom in the past, but it was caused
by deleting system fonts by mistake, whose reason no longer apply to
current case.  But it's likely that Windows fonts are involving this
issue, so we try to minimize Windows font manipulation to see if it
takes effect on the symptom.