Fix random star labels fail to display on correct position seen from Sol
We have not properly calculated transformation matrix from the universe
to the local (viewer's) CoordSys.  This issue has not been exposed since
it looks fine when we're viewing from a CoordSys having the same
orientation as the universe, but it's obvious when we're in Sol.
Non-random stars have not had this problem.