Fix a bug that velocity gets screwed when entering materialized star
We observe a phenomenon that the warp velocity vector suddenly goes to
unexpected direction when we approach a randomly generated star which
is just materialized.  This was caused by referring to old position vector
in warp calculation method, which should be updated when the destination
CoordSys is altered in materializeStar().

We are also observing strange velocity behavior when warping to Sol,
but it seems due to another cause.