pointgalaxy: Use galaxy field to obtain star density for StarEnum
We had been using hard-coded elliptical density distribution to
compute star count in a sector, but it's now unrealistic since
we can literally travel across the galaxy.

Now we should think about a way to limit generation of countless
number of star names which will be stack up in memory.