pointgalaxy: Cache pre-rendered cube image to improve performance
The number of blobs to draw the galaxy is so high that it could reach
tens of thousands.  Drawing them every frame is somewhat inefficient.
We can store rendered image to textures using framebuffer objects in
modern OpenGL implementations.  We pre-render the blobs graphics
into cube textures and draw them as a cube background to greatly
improve performance.

When the viewpoint moves far enough that cached textures do not match
the real scenery, the textures are refreshed, after waiting 2 seconds
for the viewport settles.

There's an issue that cached textures and direct rendering looks a
bit different due to blob rotation depending on viewpoint rotation.