pointgalaxy: Optimize cube background repaint distance
The cube background had checked integral indices vector in the galaxy
volume to see if the viewpoint had moved significantly, but it's not
quite optimal, because it could unnecessarily frequently repaint the
background when you are very close to the integral border.

Now it checks real number distance to determine to repaint only
if really the background has significantly moved.