pointgalaxy: Make galaxy image file selectable from script
I'd like to use a CVar, but CVar with string doesn't work.
Squirrel function is added to set this value, but it could run twice
for single file name if the application runs in server and client
mode.  I don't care this corner case for the time being because the
file name cannot be changed while the game is running.