Add Rigil Centaurus B b, actual planet candidate to space_debug.ssd
It has about the same mass as the Earth, but not likely habitable
since it's on much closer orbit to the star.  The planet's surface
is speculated to be molten rock ocean.  Its orbital elements are
not accurate.

Rigil Kentaurus B a might be better deleted since there's no such
naming convention which gives "a" to a extrasolar planet in astronomy.