Add threshold parameter to FindBrightestAstrobj
Since FindBrightestAstrobj became capable to return multiple light
sources, it may include too faint stars to the returned list.
Sometimes it may be useful, but we want to omit evaluating such
faintest light sources when rendering every frame.  So we add a
way to truncate faint stars by providing threshold.  Note that this
threshold is different from eclipseThreshold, which is used to
truncate eclipse checking calculation.