Change CoordSys::name's obsolete char-pointer type to dstring
Types of CoordSys::fullname and member of CoordSys::extranames are
also changed.  They are oldest data structures of the whole project
since the time when dstring type was not mature, but manually managing
heap memory for those strings is tough and error prone task.
We do not want to keep them forever.  Let's migrate eariler and test
longer.