Make TexSpheres reflect lights from other light sources
A TexSphere instance's absmag is derived from light sources
illuminating the area of the TexSphere's location.  The light sources
include other TexSpheres being illuminated, so multiple reflections
are also simulated, although the reflections are calculated only once
in a simulation step.  It means secondary and tertially reflections
delay several frames, but it wouldn't be noticeable.

The albedo property introduced in the previous commit is taken into
account for calculating intensity.

This reflection update costs O(n^2) of intensity evaluation with n
being the number of all TexSpheres in the universe, so it may be
suboptimal to calculate every frame.  We may think some mechanism
to cull CoordSys trees with low brightness.