vastspace: Allow standing on hooked Shipyard surface
Space Soldiers wear electromagnetic shoes to attach their feet onto
metallic surface in order to walk on it.  Hooking on a surface make
one to attach on the hooked surface.  Shooting another surface with
the hook or walking out of the face detaches from surface.

When attached, the player can control a Soldier like on the Earth's
surface, which means the mouse controls the view in lathe-like
manner.

This feature is so complex that only Shipyard support it for Soldier,
but it looks like it's a good starting point.

Also adjusted Soldier's camera view position which has not matched with
the model's head.