vastspace: Soldiers try to stand on next facet if stepped out one
Previously, if a Soldier is standing on a facet of a model and walk out
of it, he will detach from the facet even if the face is smoothly (in
terms of approximated geometry) connected to another.  Now the Soldier
will try to find another facet that looks continuing and stand on it.

Finding the new facet is done by tracing a ray downward.  The problem is
that the surface may be convex, which means the new facet may facing away
and standing on it is technically unstable.