surface: Render Apache's rotor blade more physically precisely
Helicopters control pitch directions by a swashplate.  The swashplate
will not change the angle of the rotor plane (which is the method we
have mistakenly rendered), but dynamically changes pitch of individual
rotor blades.  Now we introduce dynamic motions to alter rotor blades
angles to realize this effect.

Cockpit camera 5 (can be selected by F6) shows the pitch of a blade by
rotating along with the rotor.