Fix Squirrel not to raise floating point exception when SQUSEDOUBLE
Making Squirrel native floating point number (SQFloat) a double (through
defining SQUSEDOUBLE macro) occasionally cause floating point invalid
exception in the hash function.

I'm not sure why this doesn't happen in non-SQUSEDOUBLE build, but we need
some workaround to get it working in double precision world.  And it seems
that Linux Gcc is also free from the exception.  Probably it's due to
floating point mask (which is configured strictly in VastSpace).

This silly workaround is reinterpreting pointer to the floating point
number to a pointer to an unsigned integer of the same size.  This may
or may not appropriate for a hash function, but at least we get unique
hash for distinct numbers.