incube: Add DrawTextureCubeEx to draw sphere as inflated cube
There have been two methods to draw round astronomical objects;
DrawTextureSphere and DrawTextureSpheroid.  They were very different
implementation to render a shape, but neither one of them provide a way
to render height-varying terrain.  This branch is an attempt to create
the third type of method, DrawTextureCubeEx, to achieve it.

The astronomical body is rendered as an inflated cube; there are 6 faces
to each direction which are divided into meshes.  Vertices of divided
meshes are normalized in length to make the shape of a sphere.