Import simplex noise implementation
From:
http://webstaff.itn.liu.se/~stegu/simplexnoise/

The original author is Stefan Gustavson, and the codes are in
public domain.

I have modified the source's extension to .cpp to make it be
able to build with Visual C++ without compile option tweaks,
since the source contains declaration after statements, which
is not allowed in ANSI C but C++.

Also added are static specifiers to perm variables to prevent
duplicate symbol errors at link time.