incube: Implement very basic Level-of-Detail for DrawTextureCubeEx
This LOD determination is done per cube face (there are 6 of them), so
you won't get high detail for opposite side of the celestial body when
you walk on it.

This LOD granularity should become higher as the player apporaches the
surface to allow layered LOD tree for optimized rendering.