incube: Basic parallel processing for vertices generation
Generating vertices of a LOD patch takes some time, so doing it in the
main thread blocks real-time simulation for a moment.  We'd want to
parallelize it by performing the generation in a worker thread.

This is done by C++11's std::threads.  It works, but might be very
inefficient because a thread is created for each LOD patch.  There should
be threads of the same number as the processing cores who share jobs
collaboratively to disperse loads.

There's also an issue which causes runtime error in program exit.
It probably has something to do with std::thread's destructor.