surface: Apply gravity to LandVehicle by nearby RoundAstrobj's mass
Traditionally, the gravity is treated as uniform, straight down force in
Bullet Dynamics Engine.  Although this approximation works to some extent
in a limited range, our universe doesn't fit in there.

If there is a RoundAstrobj next to a SurfaceCS, the uniform gravity is
turned off and law of universal gravitation is applied to LandVehicles
in there.  Other types of Entities should follow.

Ceres is too lightweight to demonstrate this feature.