Properly delete the context of WGL on exit to fix the weird problem
There's a strange issue that explorer and other applications' graphics
corrupt when OpenGL application is repeatedly created and destroyed.
This issue is fairly rare that we might see it about once two weeks
without restarting the OS.  Once this happened, there are no way but
restarting the computer to fix the state.

This phenomenon is observed for other OpenGL games and the reason is not
known, but it seems to hang around since I have had Windows 7.
This specific OpenGL application (VastSpace) has been suffered from
this problem for a while and I thought the reason has something to do
with Windows font objects, but now that we observe it on other programs
which does not use them, the reason might be different.

I even bought a new graphics card to see if it changes, but it didn't.

After examining the source codes, I came to think that omitting WGL
context cleanup on program termination could be the reason.
Doing this properly may or may not help.  Several weeks will tell.

This change is back-ported from OculusBilliard's similar code fix.