incube: Introduce base level of noise for better details
Large RoundAstrobjs (such as the Earth) tend to have very little variation
in height compared to their radii.  In such a case, lower octaves in
terrain noise does not matter.  Base height can even be given as a texture,
so we want to supress lower octaves.  Specifying base level achieves it.