eng: Make a VoxelEntity be visible within reasonable distance
Our old logic to determine whether a voxel is visible is optimized
for Minecraft-like infinite field, but we prefer everything to be drawn
in space.

Also, a person in a space suit (Soldier from vastspace mod) is added to
the demo mission to show the scale of the voxels.

Currently, voxel size is set to 2.5 meters to match Space Engineer's
large block.