eng: Enable entering and controlling a VoxelEntity for Engineer
First, VoxelEntity is made inherit Autonomous in order to allow moving
like a space vehicle.  Voxel asteroids will never be controlled like an
Autonomous vehicle, but they share the base class for voxel-made space
ships.

Second, Engineer can be a controller of a VoxelEntity.  When an Engineer
enters a cockpit block in a VoxelEntity, it becomes a controller.
When controlling, all player inputs to the Engineer are redirected to
the VoxelEntity.

There's an issue that mouse movements do not take effect when controlling
a VoxelEntity, but numpad keys can be used to rotate the ship.

There's also anothor issue that thrusters, reactors, total mass and
inertia do not affect the behavior of controlled VoxelEntity.
Heavy ships should have low maneuverability, ships with many thrusters
should accelerate quickly, and so on.

Finally, multiple include guard is added to engdemo.nut because it seems
to be executed twice (thus generating two VoxelEntities and Engineers at
the same position, which causes errors).