Enable anisotropic acceleration for Autonomous descendants
Previously, Autonomous class has assumed that the acceleration is
irrelevant of directions, except that forward acceleration is twice as
high as other directions.  We want means to override this behavior
without redefining Autonomous::maneuver() in the derived class.

Adding virtual function for querying the isotropic acceleration to the
class does not fit in the scheme of using ManeuverParams as the master
of movement behavior of a ship.  So we add parameters to ManeuverParams
which defaults to isotropic.