eng: Make VoxelEntity constructor have proper predicate function for volume
Previously, we had no problem for the constructor for VoxelEntity with
a WarField as the argument, but haven't noticed that one with Game as
the argument had a problem.  The latter does not initialize VoxelEntity's
volume member variable properly, which results runtime errors when adding
more than one CellVolumes to it.

VoxelMap type needs a predicate function to compare index in 3D space
with the type Vec3i, which is provided on construction.

This issue does not show up unless we use unserialize function to restore
a VoxelMap from a stream.