eng: Fix enumerating solid cells in VoxelEntity to work properly
We have used _scanLines to enumerate solid cells in a CellVolume, but
it turned out to be incorrect.  _scanLines are used for drawing borders
(i.e. where different cell types touch each other), which is not
necessarily the same as solid cells.  To avoid this issue, we added
another member variable, solidScanLines, to store scanline information
of solid cells similar to _scanLines.

Also, usage of ranged for loop to update these scanlines were incorrect
in VoxelEntity::unserializeBlocks().  It created temporary copies of
CellVolume and updated them.