eng: Optimize VoxelEntity::updateManeuverParams()
Our usage of lambda expression and std::function was not optimal
since construction of std::function is not trivial.  We changed it so
that the construction is called only once in an invocation of
updateManeuverParams().

Also added reference declaration for the iterator in the volume loop to
prevent temporary copies of CellVolume which is 32767 bytes long.