Make Autonomous's angular acceleration take inertia into account
Previously, Autonomous::maneuver() didn't consider moment of inertia
and assumed torque the same as angular acceleration.  In reality, if
bbody is used, we pass torque to Bullet Dynamics Engine and make
it compute angular acceleration, so that assumption is invalid.

We should calculate torque to achieve desired angular acceleration,
but it's indeterminate because moment of inertia differs in axes.

We should really have a torque instead of an angular acceleration as
one of the characteristics of maneuverability, but ManeuverParams have
the acceleration for historical reasons.