Change type of strings in stellar_file.cpp into gltestp::dstring
Although gltestp::dstring and cpplib::dstring's functionality are almost
the same, we prefer gltestp::dstring over cpplib::dstring, because
cpplib::dstring cannot be dynamically linked by extension DLLs.

The StellarContext structure is not ought to be used by extension
DLLs, so it's not strictly required to replace by gltestp::dstring,
but we'd like things to be unified.