Restructure SSD parser from recursive descent to tcl-like syntax
The new parsing algorithm incorporates recursive calls, but it's not
recursive descent parser.  It just parses the whole file first and
pass verbatim string enclosed in braces to commands, which in turn
invoke another interpreter.  It's just what the tcl parser does.

Using temporary variables to store large strings is not a major problem,
but the main drawback of the new parser is that we must place a brace in
the same line of the command that the argument is passed to.
For example,

	astro "Earth" {
		...
	}

is permitted, but

	astro "Earth"
	{
		...
	}

is not.  You could also use a backslash to escape a newline for line
continuation.

This is exactly the same restriction as tcl's and should be acceptable.
In fact, accepting this restriction makes the syntax uniform for all
commands and control statements like if, else, for and while.
I must admit that some people prefer the latter coding style, though.