Very first step toward making SSD commands be dynamically defined
We have been using fixed codes for interpreting commands before falling
back to CoordSys::readFile(), but it should be inefficient and
counterintuitive to extend functionality.  We now use a hash map to resolve
commands, so it will scale well.

It's first step and merely a proof of concept with just one command,
i.e. "define".