metric: First steps toward making base length unit meters
We have been sticking to kilometers metric for literally years,
but now it's clear that it won't work well with Bullet Dynamics
Engine.  BDE seems to assume that collision objects have sizes
that having the same order of magnitude as ones, while our
coordinate systems make them thousandths.  In this condition,
certain collision shapes do not properly generate collisions
with others.  It might be something to do with collision margins,
but I'm not really sure.

The original design that use kilometers as base length unit was
decided by consideration of possibility of treating distant
solar systems that can be several millions of lightyears away,
but it's just matter of human perception; double precision
floating point numbers have sufficient exponent width (+-1000)
for space-involved computation such as square, cube, square root
or inverse.  Computed results are not affected by the base unit
as long as the significant precision is the same.  Just be sure
not to use single precision numbers in the middle of calculation.

There are so many codes that involve the spatial scale, so this
change wouldn't finish in few days, but we will start from the
viewing clip plane distances.

From now on, this batch of change is marked with "metric:"
in the first line of commit logs.